+++ packages/services/f-analytics/src/mixins/analytics.mixin.vue
@@ -6,22 +6,29 @@ import {
     COUNTRY_INFO,
     DEFAULT_APP_ID,
     DEFAULT_APP_TYPE,
-    MAP_ROUTE_TO_FEATURE_NAME,
     IDENTITY_PROVIDERS,
     GRANT_TYPES
 } from '../constants';
+import {
+    getDisplaySize,
+    getOrientation,
+    mapRouteToGroup,
+    mapRouteToFeature
+} from '../utils/helpers';
 
 export default {
     computed:
     {
         ...mapState('f-analytics', [
             'platformData',
-            'userData'
+            'userData',
+            'pageData'
         ]),
 
         isServerSide () {
             return typeof (window) === 'undefined';
         }
+
     },
 
     created () {
@@ -31,29 +38,33 @@ export default {
     methods: {
         ...mapActions('f-analytics', [
             'updatePlatformData',
-            'updateUserData'
+            'updateUserData',
+            'updatePageData'
         ]),
 
         prepareServersideAnalytics () {
             if (this.isServerSide) {
                 // Only available serverside
                 const platformData = { ...this.platformData };
+                const pageData = { ...this.pageData };
 
                 platformData.environment = process.env.justEatEnvironment || 'localhost';
                 platformData.version = process.env.FEATURE_VERSION || '0.0.0.0';
                 platformData.instancePosition = process.env.INSTANCE_POSITION || 'N/A';
-
                 // Is of type `httponly` so need to read serverside
                 platformData.jeUserPercentage = this.$cookies.get('je-user_percentage') || null;
 
+                pageData.httpStatusCode = this.$ssrContext?.res?.statusCode || 0;
+
                 this.updatePlatformData(platformData);
+                this.updatePageData(pageData);
             }
         },
 
         preparePlatformData () {
             const platformData = { ...this.platformData };
 
-            platformData.name = MAP_ROUTE_TO_FEATURE_NAME[this.$route.name] || this.$route.name;
+            platformData.name = mapRouteToFeature(this.$route.name);
             platformData.appType = DEFAULT_APP_TYPE;
             platformData.applicationId = DEFAULT_APP_ID;
             platformData.userAgent = navigator.userAgent || 'N/A';
@@ -83,11 +94,36 @@ export default {
             this.updateUserData(userData);
         },
 
+        preparePageData ({ conversationId = '', requestId = '', authToken = undefined } = {}) {
+            const pageData = {
+                ...this.pageData
+            };
+
+            pageData.group = mapRouteToGroup(this.$route.name);
+            pageData.name = mapRouteToFeature(this.$route.name);
+
+            if (pageData.name === 'Checkout 1 Overview' && !authToken) {
+                pageData.name = 'Checkout 1 Guest';
+            }
+
+            pageData.isCached = false;
+            pageData.conversationId = conversationId;
+            pageData.requestId = requestId;
+
+            if (!this.isServerSide) {
+                pageData.display = getDisplaySize();
+                pageData.orientation = getOrientation();
+            }
+
+            this.updatePageData(pageData);
+        },
+
         pushAnalytics () {
             const dataLayer = window.dataLayer || [];
             dataLayer.push({
                 platformData: { ...this.platformData },
-                userData: { ...this.userData }
+                userData: { ...this.userData },
+                pageData: { ...this.pageData }
             });
         }
     }